Configuration is now exclusivly managed by the console
