+++ vm/Cargo.toml
@@ -6,6 +6,7 @@ authors = ["Simon Whitehead <chemnova@gmail.com>"]
 [dependencies]
 rs6502 = "0.2.0"
 find_folder = "0.3.0"
+app_dirs = "1.1.1"
 rustc-serialize = "0.3.22"
 
 [dependencies.sdl2]
